perf: lazy load tools route assets
